requirements.txt
# Core LangChain and Neo4j
langchain>=0.3.8
langchain-neo4j>=0.1.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Neo4j Driver
neo4j>=5.14.0

# LLM Providers
openai>=1.50.0

# Data Processing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# API (optional)
fastapi>=0.104.0
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.10.0
ruff>=0.1.0
